R2-1→R2-14: Grip二审全部修复
R2-1: PlayMode fixture补5字段+Gate补requiredFields
R2-2: 回执发两遍→单路径发送
R2-3: 重连双轨→单轨网络线程
R2-4: UDP Gate会话重置接线
R2-5: error帧分类
R2-6: SessionMirror long→int
R2-7: UDP绑定Loopback
R2-8: 单调时钟注入
R2-9: session_id校验
R2-10: TcpClient超时
R2-11: 单一出队入口
R2-12: 时钟同步字段
R2-13: schema_version门禁
R2-14: 测试服务器UTF8修复

编译状态：Runtime+EditModeTests+PlayModeTests零错误
TASK_PLAN移出Assets，.gitignore加*.slnx

+++ 02-技术研发/04-Unity视觉/SRP-Weather-Visual/Assets/U01/Runtime/SessionMirror.cs
@@ -169,8 +169,8 @@ namespace SRP.U01
             {
                 if (evt.payload.TryGetValue("module_id", out var mid))
                     newModule = mid?.ToString();
-                if (evt.payload.TryGetValue("module_position", out var mpos) && mpos is int mp)
-                    newModulePos = mp;
+                if (evt.payload.TryGetValue("module_position", out var mpos) && mpos is long mp)
+                    newModulePos = (int)mp;
             }
             if (evt.event_type == "segment" && evt.payload != null)
             {
